src/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-dotenv==1.4.2
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
